==> notebooks/00-Data_Merge.ipynb <==
# Merging `only` and `plus` dataframes

Working with only the data from the release, we have two files:

    TEDonly_speakers_final.csv
    TEDplus_speakers_final.csv

The first thing we will do is to load both as `pandas` dataframes, but I think we also want to consider creating a merged file -- something like `TEDall.csv`. We are only going to need to do this once. From then on, we will work with either that data or an extended version of that data.

import pandas as pd

df_only = pd.read_csv('../data/Release_v0/TEDonly_speakers_final.csv')
df_plus = pd.read_csv('../data/Release_v0/TEDplus_speakers_final.csv')

print("df_only = ", df_only.shape, "\n" + "df_plus = ", df_plus.shape)

That's a lot of columns. Let's check the column names so that, first, we know we can merge the two dataframes without a problem, and, second, so that we can make sure to use the right column name later to get the texts of the talks.

# The pythonic way to open a file and get data:
with open('../data/Release_v0/TEDonly_speakers_final.csv') as f:
    cols_only = f.readline().strip().split(",")
with open('../data/Release_v0/TEDplus_speakers_final.csv') as f:
    cols_plus = f.readline().strip().split(",")

if cols_only == cols_plus:
    print("We are clear to merge.")
else:
    print("Houston, we have a problem.")

Pandas allows us to **concatenate** dataframes by row, when our observations have the same metadata; **merge** by columns, when we have additional information for our observations; and **join**, which has all the usual functionalities, and complexities, of SQL joins. Our dataframes have the same columns, so we are going to concatenate the two and then write the combined data to a new `csv`.

To do that, we are going to add a key column to each dataframe so that if we ever need to know which of the two dataframes a particular talk came from, we will be able to get that information readily. All we are going to do is add a column called `set` which will have one of two values, `only` or `plus`. The basic method of doing this is:

```python
df['Name']='abc'
```

But we are going to be a little more elaborate, because we would like to have this information on the left-hand side of the dataframe.

df_only.insert(0, 'Set', 'only')

df_only.head(3)

df_plus.insert(0, 'Set', 'plus')
df_plus.head(3)

And now to merge the two dataframes. If all goes well, we should have a dataframe with a size of `1747 x 28`.

df_all = pd.concat([df_only, df_plus])
df_all.shape

Okay. Now let's check the contents, maybe trim that "Unnamed column", and then save this to a file. (To do this, I used `df_all.head(3)` and `df_all.tail(3)` several times to check the columns and the indexing. I also used `df_all = df_all.reset_index(drop=True)` to, well, reset the index.

df_all.tail(3)

In order to remove the column which had the vestigial index, we need to see all the column names:

columns = df_all.columns.values.tolist()
print(columns)

Now that we know the exact name of the column, we can drop it:

df_all_dropped = df_all.drop('Unnamed: 0', 1)

And, of course, a quick double-check to make sure everything is as it should be:

df_all_dropped.tail(3)

And let's write this to a `csv` and get on with the analysis.

df_all_dropped.to_csv('../data/TEDall_speakers.csv', index=False)

## Loading the New Dataframe.

Loading the new dataframe.

df = pd.read_csv('../output/TEDall_speakers.csv')
df.shape

df.dtypes
