==> notebooks/01-Trends-04-Term_Year_Matrix-Build.ipynb <==
# Into the Matrix

## Notes
* [Hedonometer](https://hedonometer.org/timeseries/en_all/)

## Imports and Data

# Imports
import pandas as pd, re
from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfTransformer

# Load the Data
dfAll = pd.read_csv('../output/TEDall.csv')

We use `shape` below just to make sure the dataset loaded as expected and then `list()` to remind us of our columns: we are looking for the column that distinguishes between the main TED events and the various extra events. We are going to focus on the main events for the time being.

print(dfAll.shape, list(dfAll))

# Remind ourselves what the terms are to distinguish
# between TED main talks and all the other talks
print(set(dfAll.Set.tolist()))

# Filter the dataframe to just the TED main talks:
main = dfAll[dfAll['Set']=='only']
main.shape

# Count the number of talks per year:
main['presented'].value_counts().sort_index(ascending=True)

## Creating Pseudo-Documents for Each Year

It looks like there's not much point in including the first five years on record: they total to 11, which is only half as many as the total of 28 for 2002. 

The easiest way to proceed is:

1. Concatenate all the texts of the talks into one big pseudo-document for each year
2. Drop the first five years

We start there with concatenating all the texts of the talks into a pandas series with the years as index. In pandas you can [concatenate strings][] based on some other criteria: here we are *grouping by* the year a talk was given, which is `presented` in our dataset. (We use the `all_years` variable initially so that we can call the edited series simply `years`.)

[concatenate strings]: https://stackoverflow.com/questions/27298178/concatenate-strings-from-several-rows-using-pandas-groupby

# Concatenate all the texts of the talks into one big pseudo-document for each year
all_years = main.groupby(['presented'])['text'].apply(lambda x: ','.join(x))

# Drop the first five years
years = all_years.drop([1984, 1990, 1994, 1998, 2001])
years.head(16)

### Working with a Series

`years.values` returns an array of all the values. `years.values[0]` gives you the first item in the array/list, and as you count up the index, you walk through the pseudo-document for each year. 

Not entirely intuitively, you can get the values without explicitly calling them in a `for` loop:

```python
for item in years:
    print(item[0:20])
```

But this explicit version, which here also includes the index, is bit more clear:

for index, value in years.iteritems():
    print(f'{index}: {value[0:60]}')

## Cleaning the Year Pseudo-Documents

As a test case for later work, and without being terribly important for the current experiment, we are going to clean our texts using two functions: one to remove speakers and one to remove parentheticals.

In the cell below we create our two lists, speakers and parentheticals, and then create two separate functions and then a function to combine them. 

Our first step is to create the two lists of strings we want removed:

* `parentheticals` is from previous experiments
* `speakers` is probably unpythonic in its expression but it works

parentheticals = [ "\(laughter\)", "\(applause\)", "\(music\)", "\(video\)", 
                  "\(laughs\)", "\(applause ends\)", "\(audio\)", "\(singing\)", 
                  "\(music ends\)", "\(cheers\)", "\(cheering\)", "\(recording\)", 
                  "\(beatboxing\)", "\(audience\)", "\(guitar strum\)", 
                  "\(clicks metronome\)", "\(sighs\)", "\(guitar\)", "\(marimba sounds\)", 
                  "\(drum sounds\)" ]
len(parentheticals)

speakers = dfAll.speaker_1.tolist() + dfAll.speaker_2.tolist() + dfAll.speaker_3.tolist() + dfAll.speaker_4.tolist()
print(speakers[0:10])

def remove_parens(text):
    new_text = text
    for rgx_match in parentheticals:
        new_text = re.sub(rgx_match, ' ', new_text.lower(), flags=re.IGNORECASE)
    return new_text

def remove_speaker_names(text):
    temp_text = text
    for rgx_match in speakers:
        temp_text = re.sub(rgx_match, ' ', temp_text)
    return temp_text

def clean_text(text):
    the_text = text
    cleaned = remove_parens(remove_speaker_names(the_text))
    return cleaned

`remove_speaker_names` keeps throwing a `TypeError`.

## A Quick Word Count for Each Year

Before we go any further, let's just get a quick word count for each of our years. In this section of cells, we do the following:

# Convert our series to a dataframe to make it easier to work in place:
dfYears = years.to_frame()

# Lowercase our texts
dfYears = dfYears.apply(lambda x: x.astype(str).str.lower())

# Remove everything that isn't a word, or space
dfYears = dfYears.replace('[^\w\s\+]', '', regex = True)

# Split on spaces and then count the length of the resulting list
dfYears['word_count'] = dfYears.text.apply(lambda x: len(str(x).split(' ')))

# See the results
dfYears.head(16)

## Vectorizing the Pseudo-Documents

Then we instantiate our term frequency vectorizer and turn it loose on our pseudo-documents:

1. In creating a list from the `years` series we are returning to the texts with punctuation, which we need for our `remove_parens` function to do its job. The difference almost a thousand words in the resulting matrix.

# Countvectorizer expects a list, so we create a list
texts = [ value for index, value in years.iteritems() ]

# We are going to bring our years back to the resulting term matrix below, 
# so while we are creating lists from our series, lets grab those years
# (And yes you can create two lists from one list comprehension, but don't.)
year_labels = [ index for index, value in years.iteritems() ]

# This just checks our results
print(len(texts), texts[0][0:50], year_labels[0:5])

# The usual incantation (minus the desired speaker removal for now):
vec = CountVectorizer(preprocessor = remove_parens, min_df = 2)
word_count_vector=vec.fit_transform(texts)
word_count_vector.shape

# Create a dataframe from the resulting array
X = vec.fit_transform(texts)
term_matrix = pd.DataFrame(X.todense(), columns=vec.get_feature_names())
term_matrix.shape

term_matrix['year'] = year_labels
term_matrix.set_index('year', inplace = True)

word_df = term_matrix.transpose()
word_df.reset_index(inplace=True)
word_df = word_df.rename(columns={'index': 'term'})
word_df.head()

## Finding Terms

word_df[word_df['term']=='nuclear']

terms = ['nuclear', 'global', 'climate']

# This is a non-pandas way of getting the annual frequencies:
for item in terms:
    print(word_df[word_df['term']==item])

# And this is the pandas way
word_df[word_df.term.isin(terms)]

# Let's save this dataframe 
# ==> Commented out so re-running notebook doesn't result in new file
# word_df.to_csv('../output/term_year_matrix.csv')
