==> notebooks/01-Words-TF-Trends.ipynb <==
<h1>Table of Contents<span class="tocSkip"></span></h1>
<div class="toc"><ul class="toc-item"><li><span><a href="#Loading-the-Data" data-toc-modified-id="Loading-the-Data-1"><span class="toc-item-num">1&nbsp;&nbsp;</span>Loading the Data</a></span></li><li><span><a href="#Parsing-the-Data" data-toc-modified-id="Parsing-the-Data-2"><span class="toc-item-num">2&nbsp;&nbsp;</span>Parsing the Data</a></span></li><li><span><a href="#Working-with-the-Years" data-toc-modified-id="Working-with-the-Years-3"><span class="toc-item-num">3&nbsp;&nbsp;</span>Working with the Years</a></span></li></ul></div>

# Trends

We are going to do some re-importing of texts here by year. The first time around we are going to do the combined dataset and look for overall trends, and then we will follow that up by loading both the `only` and `plus` datasets separately to see if there are any differences worth noting. Our goal here is to see what words trend not only to learn about TED talks as a developing collection of events but it might also be possible to compare the trends glimpsed here against either trends from the BYU corpus or Google Trends itself.

import pandas as pd
import re

# from sklearn.feature_extraction.text import CountVectorizer
# from sklearn.externals import joblib

## Loading the Data

Working with only the data from the release, we have two data sets, `TEDonly` and `TEDplus` that we have previously merged into `TEDall_speakers` with an additional column indicating from which data set a given talk is taken. The first thing we will do is to load those part of `TEDall` with which we would like to work -- that is, we do not need to load all of a data set. Once we have a set loaded into a dataframe, we can use `df.columns` to see its contents like this:

```python
df = pd.read_csv('../output/TEDall_speakers.csv')
df.columns

Index(['Set', 'Talk_ID', 'public_url', 'headline', 'description', 'event',
       'duration', 'published', 'tags', 'views', 'text', 'speaker_1',
       'speaker1_occupation', 'speaker1_introduction', 'speaker1_profile',
       'speaker_2', 'speaker2_occupation', 'speaker2_introduction',
       'speaker2_profile', 'speaker_3', 'speaker3_occupation',
       'speaker3_introduction', 'speaker3_profile', 'speaker_4',
       'speaker4_occupation', 'speaker4_introduction', 'speaker4_profile'],
      dtype='object')
In [6]:
```

We can do the same for its index:

```python
df.index

RangeIndex(start=0, stop=1747, step=1)
```

But to get the column names without loading the entire data set, and in so doing save on memory use, we need to do this the old-fashioned way:

with open('../output/TEDall_speakers.csv') as f:
    columns = f.readline().strip().split(",")
print(columns)

Now we are going to load only the data we need.

df = pd.read_csv('../output/TEDall_speakers.csv', 
                    usecols=["public_url","event", "published", "text"])

df.shape

## Parsing the Data

It looks like we have two places to get a date: the event at which a talk occurred and the date on which the talk was published on the TED website. For now, let's see if we can find a way to work with the date on which the talk first occurred -- the relationship between a talk and its time is complicated by the dates sometimes being years apart, but that is something we can test later: when does a talk reveal any concurrence? 

To get all the events we can create a list and then compile to a set: 
```python
events = set(df.event.tolist())
print(len(events), events)
```
Or we can do this the **pandas** way:

events = df.event.unique().tolist()
print(len(events), events)

The events with years in them look good: we should be able to get the dates out using regex. In fact, let's build that now so that we can then filter out the events with dates in their name and get a list of the events for which we will need to assign a year. (Neither `datetime` nor `isdigit()` will work here. the former expects dates to be more systematic and the latter expects dates to be set apart from words, so **regex** it is.)

years = re.findall(r'\d+', str(events))
print(years)

In a moment, we will use this either to add a label which is only a year or to replace the label for a series with the year -- not in our original dataset, of course. 

It is probably possible to find the strings without years in them in the set above, but with only 58 items and only a half dozen, at a glance, we can do this by hand: `TEDGlobal>NYC, TEDGlobal>Geneva, TEDGlobal>London, TEDSummit, TEDGlobalLondon.`

With that list in hand, let's search the web and find the dates we need.
```
TEDGlobal>NYC: 2017
TEDGlobal>Geneva: 2015
TEDGlobal>London: 2015
TEDSummit: 2016
TEDGlobalLondon: 2015
```
So now to do a **replace** on those strings.

To replace *in situ*:

```python
df['event'].replace(
    to_replace=['ABC', 'AB'],
    value='A',
    inplace=True
)
```

To create a new column:

```python
df['elderly'] = np.where(df['age']>=50, 'yes', 'no')
```

Ack. The above code works on numbers, but we are working with strings:

```python
search = []    
for values in df['col']:
    search.append(re.search(r'\d+', values).group())

df['col1'] = search
```

search = []
for event in df['event']:
    search.append(re.search(r'\d+', str(event)).group())
    
print(search[0:20])

I'm having a hard time with doing this in place in the dataframe itself, so let's try splitting this column off as a list that we will work on and then append it back as a new column with the year. So we are going from event to year.

Here's the pseudo-code as I imagine it at this moment:

```python
for event in events:
    if replace produces a four-digit number == True then done
    else replace these things as follows:
        TEDGlobal>NYC: 2017
        TEDGlobal>Geneva: 2015
        TEDGlobal>London: 2015
        TEDSummit: 2016
        TEDGlobalLondon: 2015
```

We need to start by re-populating the `events` list with all the values and not simply the unique values.

events = df.event.tolist()
type(events[0])

Let's start with the five replacements listed above, that way we can insert the dates and those dates will be kept by the regex we use to filter out letters. To do this, we are going to use a simple dictionary that has the five events that need a date. The small test established the functionality, if not the efficiency, of the code:

test = ["TEDGlobalLondon", "TED2006", "TED 2007", "TEDGlobalLondon", "TEDSummit"]

replacements = {"TEDGlobal>NYC": "2017",
                "TEDGlobal>Geneva": "2015",
                "TEDGlobal>London": "2015",
                "TEDSummit": "2016",
                "TEDGlobalLondon": "2015"}

for key, value in replacements.items():
    test = [w.replace(key, value) for w in test]
print(test)

We can use the same dictionary, `replacements` on our master list of events:

for key, value in replacements.items():
    events = [w.replace(key, value) for w in events]

Next up, we need to apply our regex to sweep out all the characters:

# First, we need to repopulate the
events = re.findall(r'\d+', str(events))
uniques = set(events) # We're working with a list, 
                      # so we use set here instead of pandas' "unique"
print(uniques)

print(events[0:50])

Okay, it looks like we have the years, and so we'll add this back to our dataframe:

df['year'] = events

df.to_csv('../output/TEDall_years.csv')

## Working with the Years
