==> notebooks/01d-Terms-One_Doc_Wonders.ipynb <==
# One Hit Wonders

## Summary

After loading the data, we use the modified vectorizing protocol -- though it's not entirely necessary -- to create our term matrix which only allows for words that appear in one document. 

After converting the matrix into a dataframe, we sum the dataframe into a series, and then filter the series for those words that appear more than once in a document and then explore the context in which those terms appear.

## Imports and Data

# Imports
import pandas as pd, re, nltk
from sklearn.feature_extraction.text import CountVectorizer

# Load the Data
df = pd.read_csv('../output/TEDall_speakers.csv')
df.shape

urls  = df.public_url.tolist()
texts = df.text.tolist()

parentheticals = [ "\(laughter\)", "\(applause\)", "\(music\)", "\(video\)", 
                  "\(laughs\)", "\(applause ends\)", "\(audio\)", "\(singing\)", 
                  "\(music ends\)", "\(cheers\)", "\(cheering\)", "\(recording\)", 
                  "\(beatboxing\)", "\(audience\)", "\(guitar strum\)", 
                  "\(clicks metronome\)", "\(sighs\)", "\(guitar\)", "\(marimba sounds\)", 
                  "\(drum sounds\)" ]


def clean_parens(text):
    new_text = text
    for rgx_match in parentheticals:
        new_text = re.sub(rgx_match, ' ', new_text.lower(), flags=re.IGNORECASE)
    return new_text

## A Very Peculiar Vector: `max_df = 1`

vec = CountVectorizer(preprocessor = clean_parens, max_df = 1)
X = vec.fit_transform(texts)
X.shape

dfx = pd.DataFrame(X.toarray(), columns = vec.get_feature_names())

# Create a series of just the terms and the number of times they occur in total
sums = dfx.sum()

# Save to CSV
#sums.to_csv('../output/one_doc_wonders.csv')

`sums` is a pandas series. You can think of series as a one-column dataframe such that you have an index, here `sums.index` and a column with contents, here `sums.values`. You can sort by either:

sums = sums.sort_values(ascending = False)
sums.head()

## Words That Appear More Than Once But Only in One Document

Filtering a series is just like filtering a dateframe, and here we create a new series with that data and save it to a CSV.

repeated = sums [ sums > 1]
len(repeated)

# Write repeated to CSV:
# repeated.to_csv('../output/one_doc_repeated.csv')

repeated.head(10)

It's time to take see these terms in context. For now, we will convert the list of texts into a single NLTK text, but after that we will want to see in which texts these words occur.

onetext = nltk.Text(re.sub("[^a-zA-Z0-9']"," ",'\n'.join(texts)).split())

onetext.concordance("tkm")

### Initial Observations

* Unexplained (artifacts of some kind?): bf, gk, jf, tkm.
* Yup, used: telomeres (biological), Indus, fonio (some kind of grain).
* Name: Abed, Teszler.
* Song lyric: **mzuri** (in a transliterated language).
