==> notebooks/03-gender-talks.ipynb <==
# Gendering the talks 

One of our areas of interest are to see if TED talks by men and women are different. If there are differences, we would like to see if we can detect the gender of a speaker based on the words in a TED talk transcript. To pursue these avenues, we begin by "gendering" talks when possible. 

In this notebook, we use the genders of the speakers to "gender" the TED talks themselves. A talk with one speaker inherits the gender of the speaker. For a talk with two speakers, if the genders of the speakers are the same, then we proceed as if the talk had one speaker. For talks with two speakers where the genders of the speakers is not the same, we place these talks to the side. 

import pandas as pd
import csv
import string

# Load the gendered speaker file:
speakers = pd.read_csv("speakers_with_gender.csv")

ted_only = pd.read_csv('../data/Release_v0/TEDonly_final.csv')
ted_plus = pd.read_csv('../data/Release_v0/TEDplus_final.csv')

ted_only.head()

# Set the talk ID as the index and drop the unnecessary first two columns: 
ted_only = ted_only.set_index('Talk_ID')
# ted_only = ted_only.drop(columns = ['Unnamed: 0', 'Unnamed: 0.1'])

ted_only.head()

ted_only

### Slicing data for just what we need

In this notebook, we seek to combine information from our gendered-speakers file and from our ted-talks file. In the rest of this file, we will take an SQL flavored approach, that is just carrying the views of the data that we need to ultimately add the gender of our speakers to the talks that they give. 

With this view in mind, we organize our thoughts regarding our data into `PRIMARY_KEY` and `FOREIGN_KEY`:
* The `PRIMARY_KEY` for `TED_only` is the `Talk_ID`, while the four columns with `speaker_` can each be thought of the `FOREIGN_KEY`s
* The `PRIMARY_KEY` for `speakers` is the `speaker` column. Here there is no foreign key. There is also an index column that we could use as the primary key, but as we will shortly see, it will not matter which we use. 

As we proceed, we will just select the columns that are relevant to our work of gendering the talks. For `TED_only`, we will keep `Talk_ID` and the four speaker columns. For `speakers`, we will keep the name of the speaker (`speaker`), the gender (`Gender_handcheck`, labeled according to work in a previous notebook) and the index column.  

ted_slice = ted_only[["speaker_1","speaker_2","speaker_3","speaker_4"]]

ted_slice.head()

#### Dividing our talks by number of speakers

We now will divide up the `ted_slice` data frame into up to four data frames: one for each number of speakers (up to 4 speakers, which is the max number of speakers). 

##### One Speaker

First, we want to know which talks only have one speaker. To do this, we only select the rows that `NaN` for the second speaker. We will then only keep the first speaker and talk ID columns. 

talk1speak = ted_slice[ted_slice['speaker_2'].isnull()]

# https://stackoverflow.com/questions/43831539/how-to-select-rows-with-nan-in-particular-column

talk1speak.shape

# Just keep the first speaker (as the others are all NaN)

talk1 = talk1speak[["speaker_1"]]

talk1.head()


We have 966 talks that have only one speaker. 

##### Two Speakers

To figure out which talks have exactly two speakers, we select the rows that have a second speaker, but **not** a third one. We will then only keep the first two speakers and talk ID columns. 

s2temp = ted_slice[ted_slice['speaker_3'].isnull()]
talk2speak = s2temp[~s2temp['speaker_2'].isnull()]

talk2speak.shape

talk2 = talk2speak[["speaker_1","speaker_2"]]

We have 25 talks that have two speakers. 

##### Three Speakers

To figure out which talks have three speakers, we select the rows that have a third speaker, but **not** a fourth one. We will then keep the first three speaker and talk ID columns. 

s3temp = ted_slice[ted_slice['speaker_4'].isnull()]
talk3speak = s3temp[~s3temp['speaker_3'].isnull()]

talk3speak.shape

There are no talks with exactly three speakers. 

##### Four  Speakers
To figure out which talks have four speakers, we select the rows that have a fourth speaker. For this final frame we keep all four speaker columns as well as the talk ID column. 

talk4speak = ted_slice[~ted_slice['speaker_4'].isnull()]


talk4speak.shape

talk4 = talk4speak[["speaker_1","speaker_2","speaker_3","speaker_4"]]

As a quick check, we check if the total number of rows in the `talkNspeak` data frames equal the number of rows in `ted_only`:

ted_only.shape[0] == talk1speak.shape[0] + talk2speak.shape[0] + talk3speak.shape[0] + talk4speak.shape[0]

## Adding gender to the talks

Now that we have separate the talks into smaller data frames, each representing the number of speakers, we will be working to "gender" the talks. 

As with `TED_only`, we are only going to carry the columns that we need for this work, that is just the `speaker` and `gender` columns. We also have an index column, which is fine.

gender_slice = speakers[["speaker","Gender_handcheck"]]
gender_slice = gender_slice.rename(columns={"Gender_handcheck": "gender"})

# Check that we have the columns that we want, by looking at just the beginning of the data frame
gender_slice.head()

talk1test = talk1.reset_index().merge(gender_slice, left_on = "speaker_1", right_on = "speaker", how = "left")

talk1test.head()

### Gendering talks with one speaker

To gender our talks, we will use five methods from `pandas`: 
1. `reset_index()`
2. `merge()`
3. `drop()`
4. `set_index()`
5. `rename()`

For the talks with just one speaker, we will use each of these methods separately, so that it is clear what each is doing. In the later pieces, we will chain these methods together. 

First, let us remind ourselves what we are starting with:

talk1.head()

The method `reset_index()` creates a new column where the index for each row is just the row's number. We need to do this so that we don't lose the `talk_ID` column when we use `merge()` in the next step: 

talk1_step1 = talk1.reset_index()
talk1_step1.head()

# We want to take speaker gender from speakers and put it into the talks
# Remove the extra speaker column and rename the last column to "talk_gender"
talk1tmp = (talk1.reset_index()
            .merge(gender_slice, left_on = "speaker_1", right_on = "speaker", how = "left")
            .drop(columns = ['speaker']))

# We use `reset_index()` to preserve the `Talk_ID` column

talk1tmp.head()

# Reset the index back to `Talk_ID` 
talk1tmp.set_index("Talk_ID", inplace = True)

talk1tmp.head()

talk1tmp.rename(columns={"gender": "talk_gender"}, inplace = True)

talk1tmp.head()

#### Handling groups as a single speaker

In the case of an ensemble as the "only speaker," we need to take an additional step to correctly gender the talks. 

tg_male = talk1tmp["talk_gender"] == "they (group male)"

talk1tmp[tg_male]

talk1tmp.loc[tg_male,"talk_gender"] = "male"

talk1tmp[tg_male]

tg_mixed = talk1tmp["talk_gender"] == "they (group mixed)"

talk1tmp[tg_mixed]

talk1tmp.loc[tg_mixed,"talk_gender"] = "No one gender"

talk1tmp[tg_mixed]

tg_mf = talk1tmp["talk_gender"] == "they (male, female)"

talk1tmp[tg_mf]

talk1tmp.loc[tg_mf,"talk_gender"] = "No one gender"

talk1tmp[tg_mf]

### Gendering talks with two speakers

Now with that talks with just one speaker each have been "gendered," we turn our attention to the talks with multiple speakers. If the multiple speakers for a talk have the same gender, then we label the talk with that gender. 

We start with those that have two speakers. We will add the gender for the first speaker and then for the second speaker. Our procedure is similar to the above process for just one speaker. 

print(talk2.shape)

talk2.head()

We're going to make a copy of our talks with two speakers.

# We want to take speaker gender from speakers and put it into the talks
# Remove the extra speaker column and rename the last column to "gender_1"
talk2tmp = (talk2.reset_index()
            .merge(gender_slice, left_on = "speaker_1", right_on = "speaker", how = "left")
            .drop(columns = ['speaker'])
            .merge(gender_slice, left_on = "speaker_2", right_on = "speaker", how = "left")
            .drop(columns = ['speaker'])
            .rename(columns={"gender_x": "gender_1", "gender_y": "gender_2"})
            .set_index("Talk_ID"))
# We use `reset_index()` to preserve the `Talk_ID` column

# Add empty column for the talk's gender: 
talk2tmp["talk_gender"] = ""

talk2tmp.shape

talk2tmp.head()

Now we check which rows have the same genders for both speakers. We will assign those talks to have the gender of the speakers (which in this case is the same). 

talk2tmp.loc[talk2tmp["gender_1"] == talk2tmp["gender_2"], "talk_gender"] = talk2tmp["gender_1"]

Next for the rows without matching genders across the speakers, we assign the talk with the label "No one gender". We then check our resulting data slice: 

talk2tmp.loc[talk2tmp["gender_1"] != talk2tmp["gender_2"], "talk_gender"] = "No one gender"

talk2tmp

### Gendering talks with four speakers

Now with that talks with one and two speakers have been "gendered," we turn our attention to the remaining talks. As for talks with two speakers, if the multiple speakers for a talk have the same gender, then we label the talk with that gender. 

Since there was no talk with three speakers, we move right on to the only talk with four speakers. For consistency, we will use the same procedure as for talks with two speakers. 

# We want to take speaker gender from speakers and put it into the talks
# Remove the extra speaker column and rename the last column to "gender_1"
talk4tmp = (talk4.reset_index()
            .merge(gender_slice, left_on = "speaker_1", right_on = "speaker", how = "left")
            .drop(columns = ['speaker'])
            .merge(gender_slice, left_on = "speaker_2", right_on = "speaker", how = "left")
            .drop(columns = ['speaker'])
            .rename(columns={"gender_x": "gender_1", "gender_y": "gender_2"})
            .merge(gender_slice, left_on = "speaker_3", right_on = "speaker", how = "left")
            .drop(columns = ['speaker'])
            .merge(gender_slice, left_on = "speaker_4", right_on = "speaker", how = "left")
            .drop(columns = ['speaker'])
            .rename(columns={"gender_x": "gender_3", "gender_y": "gender_4"})
            .set_index("Talk_ID"))
# We use `reset_index()` to preserve the `Talk_ID` column



talk4tmp

What is interesting about this talk is that three of the speakers are male, while one is female. In keeping with our work on two speakers, we will label this as "No one gender". But for future experiments, it may be useful to remember that the majority of these speakers are male. 

# Add empty column for the talk's gender: 
talk4tmp["talk_gender"] = "No one gender"

talk4tmp

## Concatenating the Genders

Now that we have gendered all the talks in several pieces, we need to recombine these pieces. Again, in keeping with the SQL approach, as we do these combinations, we only need two columns: `Talk_ID` and `talk_gender`. 

We begin by only selecting the necessary columns. We then concatenate these pieces into one column. 

t1add = talk1tmp["talk_gender"]
t2add = talk2tmp["talk_gender"]
t4add = talk4tmp["talk_gender"]

talk_just_gender = pd.concat([t1add,t2add,t4add])

Now we merge the combined result of all the talks' genders with the main dataframe: 

ted_gender = (ted_only.drop(columns = ['Unnamed: 0', 'Unnamed: 0.1'])
              .reset_index()
              .merge(talk_just_gender, on = "Talk_ID", how = "left")
              .set_index("Talk_ID"))

ted_gender

## Splitting the talks by gender

Now we can split the talks into three groups by gender: male, female, and no one gender. By splitting our talks into these groups, we can investigate various gender dimensions of TED talk transcripts.

m_talks = ted_gender[ted_gender["talk_gender"] == "male"]
f_talks = ted_gender[ted_gender["talk_gender"] == "female"]
ng_talks = ted_gender[ted_gender["talk_gender"] == "No one gender"]

m_talks.shape

f_talks.shape

ng_talks.shape

minds = ted_gender["talk_gender"] == "male"
finds = ted_gender["talk_gender"] == "female"
ninds = ted_gender["talk_gender"] == "No one gender"


inds = (minds + finds + ninds)

sum(inds == False)

ted_gender[inds == False]

m_talks.to_csv("talks_male.csv", sep = ',')
f_talks.to_csv("talks_female.csv", sep = ',')
ng_talks.to_csv("talks_nog.csv", sep = ',')
