==> notebooks/03-Gender-Terms-NLTK-experiments.ipynb <==
# IMPORTS
import re 
import nltk
import numpy as np, pandas as pd
from sklearn.feature_extraction.text import CountVectorizer

# If needed
parentheticals = [ "\(laughter\)", "\(applause\)", "\(music\)",  
                  "\(video\)", "\(laughs\)", "\(applause ends\)", 
                  "\(audio\)", "\(singing\)", "\(music ends\)", 
                  "\(cheers\)", "\(cheering\)", "\(recording\)", 
                  "\(beatboxing\)", "\(audience\)", "\(guitar strum\)", 
                  "\(clicks metronome\)", "\(sighs\)", "\(guitar\)", 
                  "\(marimba sounds\)", "\(drum sounds\)" ]

def remove_parentheticals(text):
    global parentheticals
    new_text = text
    for rgx_match in parentheticals:
        new_text = re.sub(rgx_match, ' ', new_text.lower(), 
                          flags=re.IGNORECASE)
    return new_text

# Loading the Data in a gendered partitioned fashion: 
talks_m = pd.read_csv('talks_male.csv', index_col='Talk_ID')
talks_f = pd.read_csv('talks_female.csv', index_col='Talk_ID')
talks_nog = pd.read_csv('talks_nog.csv', index_col='Talk_ID')
talks_all = pd.concat([talks_m, talks_f, talks_nog])

# And then grabbing on the texts of the talks:
texts = talks_all.text.tolist()

print(f"From our {talks_all.shape[0]}x{talks_all.shape[1]} CSV, \
we have a list of {len(texts)} talks.")

# Tokenization

From the Terms notebook we know that tokenizers matter:

| METHOD | WORD COUNT | UNIQUE TOKENS |
| -------|------------|---------------|
| regex  |      2146  |       727     |
| nltk words  |  2576 |       725     |
| nltk wpunct | 2686  |        --     |
| scikit |      2035  |       711     |

One of the key differences between the regex and NLTK tokenizers is that the former keeps contractions together while the latter breaks them into constituent, if also somewhat odd, lexicographical units, e.g. *can't* becomes *ca* and *n't*. For now, we are going to use the regex. Later, we may find we need to switch.

For comparison, the code looks like this:

```python
# REGEX
regex = [word for word in re.sub("[^a-zA-Z']"," ", texts[0]).lower().split()]

# NLTK
w_tokens = [word.lower() for word in nltk.word_tokenize(texts[0])]
wp_tokens = [word.lower() for word in nltk.wordpunct_tokenize(texts[0])]
```

To make it easier, we are going to convert the regex line into a function.

corpus = nltk.text.Text(nltk.Text(re.sub("[^a-zA-Z0-9']"," ",'\n'.join(texts)).split()))

## Concordance

- For multi-word phrases pass a list: `text.concordance(["monstrous", "size"])`

corps.concordance(["kind", "of"])

# To save to a list:
you_know = corpus.concordance_list(["you", "know"])
you_know[5].line

print(f"you_know is a {type(you_know)} with {len(you_know)} items, each \
of which looks like this: \n\n{you_know[0]}")

# To focus on one side or the other:
corpus.concordance_list(["sort", "of"])[0].left

## Similarity

corpus.similar( "dress", num = 10 ) # Default number of words is 20.

from nltk.corpus import wordnet as wn

wn.synsets('know')

wn.synsets('know')[1].definition()

wn.synset('dress.n.01').part_meronyms()

## Common Contexts

## Collocations

texts_f = talks_f.text.tolist()
texts_m = talks_m.text.tolist()

corp_f = nltk.text.Text(nltk.Text(re.sub("[^a-zA-Z0-9']"," ",'\n'.join(texts_f)).split()))
corp_m = nltk.text.Text(nltk.Text(re.sub("[^a-zA-Z0-9']"," ",'\n'.join(texts_m)).split()))

corp_f.collocations()

corp_m.collocations()

corp_f.concordance(["sort", "of"])

corp_m.concordance(["sort", "of"])
