==> notebooks/03-Gender.ipynb <==
# Gender and TED talks

One avenue of research in this project examines what -- if any -- relationship exists between the gender of a TED speaker and the talk that they give. To explore this line of thinking, we would need to have access to the gender that each speaker identifies with. 

One limitation of the data on the TED webpages is that there is no demographic information about the speakers such as age, race, or gender. In this notebook, we craft a method for detecting speakers' genders from the existing data within the collated speaker information from the TED website, including `speaker_occupation`, `speaker_introduction`, and `speaker_profile`. 

There are two parts to this method: 1) an automated gender detector, and 2) a non-automated procedure. For the first step, we extract the number and kinds of pronouns used in each of the three speaker variables: `speaker_occupation`, `speaker_introduction`, and `speaker_profile`. We then take a conservative approach to guessing a speaker's gender where the vast majority of pronouns need to be either stereotypically male or female. For any person who do not meet such a condition, we then work to determine their self-identified gender by reading their speaker information and referencing other media about the person. For anyone that we needed additional resources, they are noted in the resulting gender file. 


## Importing needed modules and data 

import pandas as pd
import csv
import string

df_only = pd.read_csv('../data/Release_v0/TEDonly_speakers_final.csv')
df_plus = pd.read_csv('../data/Release_v0/TEDplus_speakers_final.csv')

Now that we have imported our data files, we have to extract just the data about the speakers. Noting that the speaker information comes after the talk information, we start by finding the first column related to `speaker_1`.

col_names = df_only.columns
col_list = list(col_names) 
sind = col_list.index("speaker_1")

# https://stackoverflow.com/questions/9542738/python-find-in-list

Now we remove all the columns affiliated to the talk information. 

speaker_info = df_only
speaker_info.drop(speaker_info.columns[0:sind], axis=1, inplace=True)

Now we extract information for each of the potentially four speakers per talk. We will rename the columns in each of these four data frames to be identical. We will also remove any rows that are completely empty and check for duplicates before merging all of the pieces together in one cohesive list of all the speakers. 

speak1 = speaker_info.iloc[:,0:4]
speak1.columns=['speaker', 'occupation', 'introduction', 'profile']
print(speak1.shape)
speak1 = speak1.dropna(how='all')
speak1 = speak1.drop_duplicates()
print(speak1.shape)

speak2 = speaker_info.iloc[:,4:8]
speak2.columns=['speaker', 'occupation', 'introduction', 'profile']
print(speak2.shape)
speak2 = speak2.dropna(how='all')
print(speak2.shape)
speak2 = speak2.drop_duplicates()
print(speak2.shape)

speak3 = speaker_info.iloc[:,8:12]
speak3.columns=['speaker', 'occupation', 'introduction', 'profile']
print(speak3.shape)
speak3 = speak3.dropna(how='all')
print(speak3.shape)

speak4 = speaker_info.iloc[:,12:16]
speak4.columns=['speaker', 'occupation', 'introduction', 'profile']
print(speak4.shape)
speak4 = speak4.dropna(how='all')
print(speak4.shape)

# Above from https://stackoverflow.com/questions/47060980/...
#        renaming-the-column-names-of-pandas-dataframe-is-not-working-as-expected-pytho

At this point, we have 4 smaller data frames corresponding to the four possible speakers for each talk. Now we will put them back together as one data frame and adjust the indices. 

*Note* - In this data frame the indices are the row number. To advoid creating "duplicate indices", we will have a few instances below where we use flags from `pandas` to reset the indices to match the row numbers exactly. 

speak_all = pd.concat([speak1, speak2,speak3, speak4], ignore_index = True) # <-- Sarah M. Brown

Even though we dropped the duplicates from each of the smaller data frames, this does not prevent a speaker from showing up as a first speaker in one talk and the third speaker in another talk. So again, we have to check for and remove any repeats.

# Example of remaining double
speak_all[speak_all["speaker"] =="Rhiannon Giddens"]

speak_all.drop_duplicates(inplace=True)
speak_all.reset_index(drop=True,inplace=True)

# drop_duplicates() from 
# http://www.datasciencemadesimple.com/get-unique-values-rows-dataframe-python-pandas/

# Extra flags from Sarah M. Brown

speak_all.shape

speak_all = speak_all[~speak_all.duplicated(['speaker', 'occupation', 'introduction'])]
# From https://stackoverflow.com/questions/40059994/pandas-get-a-list-of-index-from-dataframe-loc

speak_all.shape

speak_all.reset_index(drop=True,inplace=True)

# Checking that we only have one line for "Bill Gates":
speak_all[speak_all["speaker"] == "Bill Gates"]

## Distilling pronouns from prose

The challenge with this extraction is balancing efficiency (and automation) with not incorrectly identifying any speaker's gender. To this, we classify speakers in two steps. First we use an automatic classifier that very conservatively assigns binary genders using pronoun ratios. The second step involves hand checking the information on each speaker who did not receive a classification from the previous step and -- if needed -- checking additional media about the speaker.

### Step 1 - Automatic gender detection

In this step, we attempt to automatically detect the self-identification of speakers by finding pronouns within each speaker's variables: `speaker_occupation`, `speaker_introduction`, and `speaker_profile`. In this step, we compute the ratios of 1) male pronouns compared to all pronouns, 2) female pronouns compared to all pronouns, and 3) non-binary pronouns compared to all pronouns. If either of the first two ratios are very high, then we classify the speaker as male or female, respectively. We are very conservative in applying these labels, as we do not want to mis-identify any speakers. To aid in the next step (and for further investigations regarding the impact of gender), we store the counts for 1) male pronouns, 2) female pronouns, and 3) non-binary pronouns. 

### Step 1.0 - Prepare the data frame 

We start by adding the columns that we intend on filling in. So we will be adding the columns `Gender_auto`, `MaleScore`, `FemaleScore`, `NonBinaryScore` to our data frame. 

header_list = ['speaker', 'occupation', 'introduction', 'profile', 
               'Gender_auto', 'MaleScore', 'FemaleScore', 'NonBinaryScore']

speak_all = speak_all.reindex(columns = header_list)
speak_all.head()


#https://stackoverflow.com/questions/16327055/how-to-add-an-empty-column-to-a-dataframe

### Step 1.1 Create necessary systems

Create the _global_ lists of pronouns and the function for the conservative assigning of genders. 

# Pronoun lists
male_pronouns = {'he', 'him', 'his', 'himself'}
female_pronouns = {'she', 'her', 'hers', 'herself'}

nonbinary_pronouns = {'they', 'them', 'their', 'theirs', 'themself', 
                    'e', 'ey', 'em', 'eir', 'eirs', 'eirself', 
                    'fae', 'faer', 'faers', 'faerself', 
                    'per', 'pers', 'perself',
                    've', 'ver', 'vis', 'verself',
                    'xe', 'xem', 'xyr', 'xyrs', 'xemself',
                    'ze', 'zie', 'hir', 'hirs', 'hirself', 
                    'sie', 'zir', 'zis', 'zim', 'zieself', 
                    'emself', 'tey', 'ter', 'tem', 'ters', 'terself'} 

We next define a function that finds and counts the various pronouns used in the text of the `speaker_occupation`, `speaker_introduction`, and `speaker_profile` variables. 

def find_gender(input_description):
	global male_pronouns, female_pronouns, nonbinary_pronouns

	# Initialize score variables
	male_score = 0
	female_score = 0
	nonbinary_score = 0

	# Lower and isolate everyword of the description
	des_lst = input_description.lower().split()
	for word in des_lst:
		cleanword = word.strip(string.punctuation)

		# Add to the appropriate score
		if cleanword in male_pronouns:
			male_score = male_score + 1
		elif cleanword in female_pronouns:
			female_score = female_score + 1
		elif cleanword in nonbinary_pronouns:
			nonbinary_score = nonbinary_score + 1

	total = male_score + female_score + nonbinary_score

	if total == 0: # Only happens if there are no pronouns
		gender = 'no pronouns'
	# elif (nonbinary_score <= (.1)*total):
	# Note: The above line is too harsh. 
	
	# If there are none of two kinds of pronouns, then set the automatic gender
    #     to the third one from the set of {female, male, non-binary}
	elif (male_score == 0) and (female_score == 0):
		gender = 'non-binary'
	elif (male_score == 0) and (nonbinary_score == 0):
		gender = 'female'
	elif (female_score == 0) and (nonbinary_score == 0):
		gender = 'male'

	# If there only one of the set {female, male, non-binary} does not appear
    #     then see if there is a strong lean towards one of the remaining two
    #     in the set. 
	elif (nonbinary_score <= 1):
		score = (female_score - male_score) / (female_score + male_score)
		if score > 0.3:
			gender = 'female'
		elif score < -0.3:
			gender = 'male'
		else:
			gender = 'undetected'
	elif (male_score == 0):
		score = (female_score - nonbinary_score) / (female_score + nonbinary_score)
		if score > 0.3:
			gender = 'female'
		elif score < -0.3:
			gender = 'non-binary'
		else:
			gender = 'undetected'
	elif (female_score == 0):
		score = (male_score - nonbinary_score) / (male_score + nonbinary_score)
		if score > 0.3:
			gender = 'male'
		elif score < -0.3:
			gender = 'non-binary'
		else:
			gender = 'undetected'
	else:
		gender = 'last case'

	return (gender, male_score, female_score, nonbinary_score)

Now we loop over each row, computing the number of times each kind of pronoun comes up and use the above detector to conservatively assign gender groups. 

for row in speak_all.itertuples():
    inds = row.Index
    des_list = [str(row.occupation), str(row.introduction), row.profile]
    descript = " ".join(des_list)
    #auto_gender, male_score, female_score, nonbinary_score = find_gender(descript)
    speak_all.loc[inds,['Gender_auto', 'MaleScore', 'FemaleScore', 'NonBinaryScore']] = find_gender(descript)

# https://cmdlinetips.com/2018/12/how-to-loop-through-pandas-rows-or-how-to-iterate-over-pandas-rows/
# https://www.programiz.com/python-programming/methods/string/join

('male', 5, 0, 0)
('male', 5, 0, 0)
('female', 1, 11, 0)
('male', 8, 0, 2)
('male', 13, 0, 2)

results = speak_all['Gender_auto'].value_counts()

#https://stackoverflow.com/questions/38309729/count-unique-values-with-pandas-per-groups/38309823

results

speak_all.loc[speak_all["Gender_auto"] == 'undetected']

## Step 2 - Hand checking

We now turn to hand checking anyone who was not assigned a binary gender in the automatic detector. Before doing that handchecking, we create a new column in our data frame `Gender_handcheck` where we will record the handchecked genders. 

header_list = ['speaker', 'occupation', 'introduction', 'profile', 
               'Gender_auto', 'MaleScore', 'FemaleScore', 'NonBinaryScore', 'Gender_handcheck']

speak_all = speak_all.reindex(columns = header_list)
speak_all.head()

If a person _was_ assigned a binary gender in by the automatic detector then we copy the assignment into `Gender_handcheck`

speak_all.loc[speak_all["Gender_auto"] == 'male','Gender_handcheck'] = "male"
speak_all.loc[speak_all["Gender_auto"] == 'female','Gender_handcheck'] = "female"

speak_all.to_csv("gender_step1.csv", sep = ',', index=False)

### Hand evaluation

As a human, it iscan be straight-forward to read a sentence or two from a TED presenter's bio and determine their gender. So for any speaker who was not given a binary gender label from the conservative auto-labeling function above, we read their bio and add a human label for their gender. 

This was first done for those whose gender was `undetected`. Then we will repeat the process on those labeled as `non-binary`, `no pronouns`, and `last case`. 

First, we begin by isolating the rows in `speak_all` where the automatic gender (encoded in `Gender_auto`) is `undetected`. We then save this dataset for handchecking. 

speak_ud = speak_all.loc[speak_all["Gender_auto"] == 'undetected']
speak_ud.to_csv("gender_undected.csv", sep = ',', index=False)

#### Completing the first round of handchecks

After saving the rows labeled as `undetected`, we handcheck the data by reading the information in the associated `introduction` and `profile` features. When necessary, we explore additional resources like wikipedia and personal websites to find self-identified genders of the speakers. 



#### Completing the second round of handchecks

Once this first phase of handchecking is done, we will continue with the next round of handchecking for those labeled as `non-binary`, `no pronouns`, and `last case`. To begin, we now create a spreadsheet to repeat the above handchecks on those automatically labeled as `non-binary`, `no pronouns`, and `last case`. 

speak_nb = speak_all.loc[speak_all["Gender_auto"] == 'non-binary']
speak_np = speak_all.loc[speak_all["Gender_auto"] == 'no pronouns']
speak_lc = speak_all.loc[speak_all["Gender_auto"] == 'last case']

speak_all_checks = pd.concat([speak_nb, speak_np, speak_lc], ignore_index = True)

speak_all_checks

speak_all_checks.to_csv("gender_check2.csv", sep = ',', index=False)

### Creating one complete speaker file

In the previous section, we created two files that needed to be handchecked. For completion and clarity, we created two files `gender_check2_checked.csv` and `gender_undected_checked.csv` that contain all the fully checked speakers.

In this section, we will create one file that contains all the speakers with their handchecked genders included. Specifically, we are adding the handchecked genders from `gender_check2_checked.csv` and `gender_undected_checked.csv` to `gender_step1.csv`

# Load gender_step1.csv
speak_all = pd.read_csv("gender_step1.csv", sep = ",")

# Load gender_undected_handchecked.csv, gender_check2_checked.csv and gender_undected_checked.csv 
handcheck1 = pd.read_csv("gender_undected_checked.csv", sep = ",")
handcheck2 = pd.read_csv("gender_check2_checked.csv", sep = ",")

# Merge based on name and occupation. Use loc? 
# https://stackoverflow.com/questions/49928463/python-pandas-update-a-dataframe-value-from-another-dataframe

# Trying code from here
# https://stackoverflow.com/questions/49928463/python-pandas-update-a-dataframe-value-from-another-dataframe
# pd.concat([df1,df2]).drop_duplicates(['Code','Name'],keep='last').sort_values('Code')

speaker_merge = pd.concat([speak_all, handcheck1, handcheck2])
speaker_merge = speaker_merge.drop_duplicates(["speaker","occupation"],keep='last')

We use the flag `keep = 'last'` because we stack both of the handchecked files under the results from the automatic checker. So the speakers in the handcheck files do not have an automatically assigned gender in the `speak_all` dataframe. 

# check that we have not added extra speakers nor removed any

print(speak_all.shape == speaker_merge.shape)

#### Last step

Finally, save the results of the merged gender assignments:

speaker_merge.to_csv("speakers_with_gender.csv", sep = ',', index=False)
