==> notebooks/04-Lexicalities-Tense.ipynb <==
# Tense

After we load our modules and data, creating a list from the text column so that we can work with subsets of texts, we explore part-of-speech tagging and how to focus on the verbs and their tenses and then count those.

import nltk, re, pandas as pd
from collections import Counter

df = pd.read_csv('../output/TEDall_speakers.csv')
texts = df.text.tolist()

## Working through POS-tagging

The function below is just so we can pull some sentences from various TED talks in order to test our POS-tagging script as we develop it.

def sentience (the_string):
    sentences = [
            [word.lower() for word in nltk.word_tokenize(sentence)]
            for sentence in nltk.sent_tokenize(the_string)
        ]
    return sentences

The first thing we are going to do is grab two sentences from a talk and use them as our `test`. There has to be a more elegant, or even pythonic, way to do this, but nothing I tried with nested list comprehensions of itertool's `chain` worked. This did:

test = []
for i in range(len(text2[0:2])): #Traversing through the main list
    for j in range (len(text2[i])): #Traversing through each sublist
        test.append(text2[i][j]) 

print(test)

print(nltk.pos_tag(test))

## A Toy Example

tenses = '''
Molly wanted to leave.
Molly wants to leave.
Molly will want to leave.
'''
print(nltk.pos_tag([word.lower() for word in nltk.word_tokenize(tenses)]))

The only thing in which we are interested in the current moment is the tense of the main verb. NLTK has the following possibilities:


| Tag | Object                          | Example |
|--:--|:--------------------------------|:--------|
| VB  | verb, base form                 | take    |
| VBD | verb, past tense                | took    |
| VBG | verb, gerund/present participle | taking  |
| VBN | verb, past participle           | taken   |
| VBP | verb, sing. present, non-3d     | take    |
| VBZ | verb, 3rd person sing. present  | takes   |

We are looking for: present tense verbs (VB, VBP, VBZ) and past tense verbs (VBD). We will ignore the complexities introduced via participles. And we will have to explore how to track future (will) and, possibly, subjunctive tenses (would, could, should). 

So, what the NLTK POS-tagger returns is a list of tuples with punctuation still in place, so we need to find the tuples with the form `('word', 'tag')` where `tag == ` one of our four possibilities (VB, VBP, VBZ for present tense and VBD for past tense)

def sentenser(sentence_words):
    fps_length=len(first_singular_words.intersection(sentence_words))
    fpp_length=len(first_plural_words.intersection(sentence_words))
    sec_length=len(second_words.intersection(sentence_words))
    third_length=len(third_person_words.intersection(sentence_words))
    

    if fps_length > (fpp_length + sec_length + third_length):
        perspective = 'first person singular'
    elif fpp_length > (fps_length + sec_length + third_length): 
        perspective = 'first person plural'
    elif sec_length > (fps_length + fpp_length + third_length): 
        perspective = 'second person'
    elif third_length > (fps_length + fpp_length + sec_length): 
        perspective = 'third person'
    else:
        perspective = 'mixed'
    return perspective

def count_tense(sentences):
    
    """ REQUIRES: from collections import Counter"""
        
    sents = Counter()
    words = Counter()
    
    for sentence in sentences:
        perspective = perse_sentence(sentence)
        sents[perspective] += 1
        words[perspective] +=len(sentence)
        
    return sents, words

def parper(text):
    
        """ REQUIRES: import nltk """
        
        sentences = [
            [word.lower() for word in nltk.word_tokenize(sentence)]
            for sentence in nltk.sent_tokenize(text)
        ]
        
        sents, words = count_perspective(sentences)
        total = sum(words.values())
        
        for perspective, count in words.items():
            pcent = (count / total) * 100
            nsents = sents[perspective]
            
            print(f"{pcent:.2f} {perspective} ( {nsents} ) ")

The example below simply establishes that gender is not straightforward: there is no guarantee that someone using gendered words, as constructed above, is actually talking about women.

parper(words)

Al Gore's text is a bellweather because it sits at `0` in the list index. Here we mix it up with Majora Carter's talk on urban renewal:

parper(texts[2])
